@@ -17,12 +17,17 @@
             "FriendsOfPhpSpec\\PhpSpec\\Expect\\": "src/"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "FriendsOfPhpSpec\\PhpSpec\\Tests\\": "tests/"
+        }
+    },
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4",
         "phpspec/phpspec": "^6.0 || ^7.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0 || ^8.0 || ^9.0",
+        "phpunit/phpunit": "^9.0 || ^10.0",
         "squizlabs/php_codesniffer": "^3.5"
     },
     "scripts": {
